refactor sections and add pic
