+++ public/blue-hills.jpg
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:e80a144667bd0837d562f003cf42be808b97441324127a7ea839cec7a8e17a97
-size 5309062
+oid sha256:7dc266599296d19908105306771e4550d9ee7326793a14243aed370da251d5f0
+size 1139387
