+++ src/timeline-section/PersonalTimeline.tsx
@@ -6,12 +6,6 @@ import TimelineContent from '@material-ui/lab/TimelineContent';
 import TimelineDot from '@material-ui/lab/TimelineDot';
 import React, { Dispatch, FC, SetStateAction } from 'react';
 
-// const useLocalClasses = makeStyles(() => {
-//     return {
-//         timeline: {},
-//     };
-// });
-
 export interface PersonalTimelineProps {
     selectedTimelineItem: number;
     setSelectedTimelineItem: Dispatch<SetStateAction<number>>;
@@ -21,13 +15,6 @@ const PersonalTimeline: FC<PersonalTimelineProps> = (): JSX.Element => {
     return (
         <React.Fragment>
             <Timeline>
-                <TimelineItem>
-                    <TimelineSeparator>
-                        <TimelineDot />
-                        <TimelineConnector />
-                    </TimelineSeparator>
-                    <TimelineContent>Born</TimelineContent>
-                </TimelineItem>
                 <TimelineItem>
                     <TimelineSeparator>
                         <TimelineDot />
@@ -43,6 +30,25 @@ const PersonalTimeline: FC<PersonalTimelineProps> = (): JSX.Element => {
                 </TimelineItem>
             </Timeline>
         </React.Fragment>
+    );
+};
+
+export interface PersonalTimelineItemProps {
+    index: number;
+    onMouseOver: () => void;
+}
+
+const PersonalTimelineItem: FC<PersonalTimelineItemProps> = (): JSX.Element => {
+    return (
+        <div>
+            <TimelineItem>
+                <TimelineSeparator>
+                    <TimelineDot />
+                    <TimelineConnector />
+                </TimelineSeparator>
+                <TimelineContent>Born</TimelineContent>
+            </TimelineItem>
+        </div>
     );
 };
